chore: product 구조 분리
