+++ src/main/java/com/kakaotech/ott/ott/aiImage/presentation/dto/response/AiImageAndProductResponseDto.java
@@ -1,11 +1,11 @@
 package com.kakaotech.ott.ott.aiImage.presentation.dto.response;
 
+import com.kakaotech.ott.ott.product.presentation.dto.response.ProductResponseDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Getter
