+++ src/main/java/com/kakaotech/ott/ott/global/exception/ErrorCode.java
@@ -18,8 +18,7 @@ public enum ErrorCode {
     REFRESH_TOKEN_INVALID_SIGNATURE(HttpStatus.UNAUTHORIZED, "Refresh Token 서명이 유효하지 않습니다."),
 
     USER_DELETED(HttpStatus.FORBIDDEN, "탈퇴한 사용자입니다."),
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류가 발생했습니다."),
-    KAKAO_API_ERROR(HttpStatus.BAD_GATEWAY, "카카오 서버와 통신에 실패했습니다.");
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "서버 내부 오류가 발생했습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
