+++ src/main/java/com/kakaotech/ott/ott/home/presentation/dto/response/MainResponseDto.java
@@ -1,0 +1,4 @@
+package com.kakaotech.ott.ott.home.presentation.dto.response;
+
+public class MainResponseDto {
+}
