+++ src/main/java/com/kakaotech/ott/ott/product/application/serviceImpl/ProductDomainServiceImpl.java
@@ -1,19 +1,19 @@
-package com.kakaotech.ott.ott.aiImage.application.serviceImpl;
+package com.kakaotech.ott.ott.product.application.serviceImpl;
 
-import com.kakaotech.ott.ott.aiImage.application.service.ProductDomainService;
+import com.kakaotech.ott.ott.product.application.service.ProductDomainService;
 import com.kakaotech.ott.ott.aiImage.domain.model.AiImage;
-import com.kakaotech.ott.ott.aiImage.domain.model.DeskProduct;
-import com.kakaotech.ott.ott.aiImage.domain.model.ProductMainCategory;
-import com.kakaotech.ott.ott.aiImage.domain.model.ProductSubCategory;
+import com.kakaotech.ott.ott.product.domain.model.DeskProduct;
+import com.kakaotech.ott.ott.product.domain.model.ProductMainCategory;
+import com.kakaotech.ott.ott.product.domain.model.ProductSubCategory;
 import com.kakaotech.ott.ott.aiImage.domain.repository.AiImageRepository;
-import com.kakaotech.ott.ott.aiImage.domain.repository.DeskProductRepository;
-import com.kakaotech.ott.ott.aiImage.domain.repository.ProductMainCategoryRepository;
-import com.kakaotech.ott.ott.aiImage.domain.repository.ProductSubCategoryRepository;
+import com.kakaotech.ott.ott.product.domain.repository.DeskProductRepository;
+import com.kakaotech.ott.ott.product.domain.repository.ProductMainCategoryRepository;
+import com.kakaotech.ott.ott.product.domain.repository.ProductSubCategoryRepository;
 import com.kakaotech.ott.ott.aiImage.infrastructure.entity.AiImageEntity;
-import com.kakaotech.ott.ott.aiImage.infrastructure.entity.ProductMainCategoryEntity;
-import com.kakaotech.ott.ott.aiImage.infrastructure.entity.ProductSubCategoryEntity;
+import com.kakaotech.ott.ott.product.infrastructure.entity.ProductMainCategoryEntity;
+import com.kakaotech.ott.ott.product.infrastructure.entity.ProductSubCategoryEntity;
 import com.kakaotech.ott.ott.aiImage.presentation.dto.request.AiImageAndProductRequestDto;
-import com.kakaotech.ott.ott.aiImage.presentation.dto.request.ProductDetailRequestDto;
+import com.kakaotech.ott.ott.product.presentation.dto.request.ProductDetailRequestDto;
 import com.kakaotech.ott.ott.user.domain.model.User;
 import com.kakaotech.ott.ott.user.domain.repository.UserRepository;
 import jakarta.persistence.EntityNotFoundException;
