+++ src/main/java/com/kakaotech/ott/ott/product/domain/repository/ProductMainCategoryJpaRepository.java
@@ -1,6 +1,6 @@
-package com.kakaotech.ott.ott.aiImage.domain.repository;
+package com.kakaotech.ott.ott.product.domain.repository;
 
-import com.kakaotech.ott.ott.aiImage.infrastructure.entity.ProductMainCategoryEntity;
+import com.kakaotech.ott.ott.product.infrastructure.entity.ProductMainCategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
