+++ src/main/java/com/kakaotech/ott/ott/product/domain/repository/ProductMainCategoryRepository.java
@@ -1,0 +1,13 @@
+package com.kakaotech.ott.ott.product.domain.repository;
+
+import com.kakaotech.ott.ott.product.domain.model.ProductMainCategory;
+import com.kakaotech.ott.ott.product.infrastructure.entity.ProductMainCategoryEntity;
+
+import java.util.Optional;
+
+public interface ProductMainCategoryRepository {
+
+    ProductMainCategoryEntity save(ProductMainCategory productMainCategory);
+
+    Optional<ProductMainCategoryEntity> findByName(String mainCategoryName);
+}
