+++ src/main/java/com/kakaotech/ott/ott/product/infrastructure/serviceImpl/DeskProductRepositoryImpl.java
@@ -1,11 +1,11 @@
-package com.kakaotech.ott.ott.aiImage.infrastructure.repositoryImpl;
+package com.kakaotech.ott.ott.product.infrastructure.serviceImpl;
 
-import com.kakaotech.ott.ott.aiImage.domain.model.DeskProduct;
-import com.kakaotech.ott.ott.aiImage.domain.repository.DeskProductJpaRepository;
-import com.kakaotech.ott.ott.aiImage.domain.repository.DeskProductRepository;
+import com.kakaotech.ott.ott.product.domain.model.DeskProduct;
+import com.kakaotech.ott.ott.product.domain.repository.DeskProductJpaRepository;
+import com.kakaotech.ott.ott.product.domain.repository.DeskProductRepository;
 import com.kakaotech.ott.ott.aiImage.infrastructure.entity.AiImageEntity;
-import com.kakaotech.ott.ott.aiImage.infrastructure.entity.DeskProductEntity;
-import com.kakaotech.ott.ott.aiImage.infrastructure.entity.ProductSubCategoryEntity;
+import com.kakaotech.ott.ott.product.infrastructure.entity.DeskProductEntity;
+import com.kakaotech.ott.ott.product.infrastructure.entity.ProductSubCategoryEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
